+++ magpie/api/__init__.py
@@ -14,4 +14,4 @@ def includeme(config):
     config.add_notfound_view(not_found)
     config.add_exception_view(internal_server_error)
     config.add_forbidden_view(unauthorized_access)
-    #config.scan()
+    # config.scan()