+++ magpie/api/login/__init__.py
@@ -1,4 +1,4 @@
-from magpie.api.api_rest_schemas import *
+from magpie.api import api_rest_schemas as s
 import logging
 logger = logging.getLogger(__name__)
 
@@ -6,8 +6,8 @@ logger = logging.getLogger(__name__)
 def includeme(config):
     logger.info('Adding api login ...')
     # Add all the rest api routes
-    config.add_route(**service_api_route_info(SessionAPI))
-    config.add_route(**service_api_route_info(SigninAPI))
-    config.add_route(**service_api_route_info(ProvidersAPI))
-    config.add_route(**service_api_route_info(ProviderSigninAPI))
+    config.add_route(**s.service_api_route_info(s.SessionAPI))
+    config.add_route(**s.service_api_route_info(s.SigninAPI))
+    config.add_route(**s.service_api_route_info(s.ProvidersAPI))
+    config.add_route(**s.service_api_route_info(s.ProviderSigninAPI))
     config.scan()