+++ HISTORY.rst
@@ -3,14 +3,18 @@
 History
 =======
 
+0.9.x
+---------------------
+
+* add permissions config to auto-generate user/group rules on startup
+* fix many invalid or erroneous swagger specifications
+* attempt db creation on first migration if not existing
+
 0.8.x
 ---------------------
 
 * update MagpieAdapter to match process store changes
 * provide user ID on API routes returning user info
-* attempt db creation on first migration if not existing
-* add permissions config to auto-generate user/group rules on startup
-* fix many invalid or erroneous swagger specifications
 
 0.7.x
 ---------------------